update: routes changed and working
