+++ src/routes/chat/$chatId/index.lazy.tsx
@@ -1,39 +1,20 @@
-import { createLazyFileRoute } from "@tanstack/react-router";
-import { ChatMessages } from "@/components/chat/messages";
-import { ChatInput } from "@/components/chat/input/index";
-import { AppSidebar } from "@/components/app-sidebar";
-import {
-  ResizableHandle,
-  ResizablePanel,
-  ResizablePanelGroup,
-} from "@/components/ui/resizable";
-import { TopNav } from "@/components/topnav";
-import { Panel } from "@/components/chat/panels";
-import { DocumentDialog } from "@/components/document-dialog";
-import { CreateProjectDialog } from "@/components/create-project-dialog";
-import { useAtomValue, useSetAtom } from "jotai";
-import { resizePanelOpenAtom, selectedArtifactAtom } from "@/store/chatStore";
-import { useEffect } from "react";
+import { createLazyFileRoute, useParams } from "@tanstack/react-router";
 import type { Id } from "../../../../convex/_generated/dataModel";
-import { Navigate, useParams } from "@tanstack/react-router";
-import { motion, AnimatePresence } from "motion/react";
-import {
-  slideInFromRight,
-  slideInFromLeft,
-  layoutTransition,
-} from "@/lib/motion";
-import { Authenticated, Unauthenticated } from "convex/react";
+import { useEffect } from "react";
+import { useSetAtom } from "jotai";
+import { selectedArtifactAtom } from "@/store/chatStore";
+import { motion } from "motion/react";
+import { layoutTransition } from "@/lib/motion";
+import { ChatMessages } from "@/components/chat/messages";
+import { ChatInput } from "@/components/chat/input";
 
 export const Route = createLazyFileRoute("/chat/$chatId/")({
   component: RouteComponent,
 });
 
 function RouteComponent() {
-  const params = useParams({
-    from: "/chat/$chatId/",
-  });
+  const params = useParams({ strict: false });
   const chatId = params.chatId as Id<"chats">;
-  const resizePanelOpen = useAtomValue(resizePanelOpenAtom);
   const setSelectedArtifact = useSetAtom(selectedArtifactAtom);
 
   useEffect(() => {
@@ -42,60 +23,16 @@ function RouteComponent() {
 
   return (
     <>
-      <Unauthenticated>
-        <Navigate to="/" />
-      </Unauthenticated>
-      <Authenticated>
-        <TopNav />
-        <motion.div
-          variants={slideInFromLeft}
-          initial="initial"
-          animate="animate"
-          transition={layoutTransition}
-        >
-          <AppSidebar />
-        </motion.div>
-        <ResizablePanelGroup direction="horizontal">
-          <ResizablePanel className="flex flex-col gap-1 p-2 pt-4">
-            <motion.div
-              className="flex-1 min-h-0"
-              layout
-              transition={layoutTransition}
-            >
-              <ChatMessages />
-            </motion.div>
-            <motion.div
-              className="flex-none"
-              layout
-              transition={layoutTransition}
-            >
-              <ChatInput />
-            </motion.div>
-          </ResizablePanel>
-          <AnimatePresence mode="wait">
-            {resizePanelOpen && (
-              <>
-                <ResizableHandle />
-                <ResizablePanel defaultSize={40} minSize={25} maxSize={60}>
-                  <motion.div
-                    variants={slideInFromRight}
-                    initial="initial"
-                    animate="animate"
-                    exit="exit"
-                    transition={layoutTransition}
-                    className="h-full"
-                  >
-                    <Panel />
-                  </motion.div>
-                </ResizablePanel>
-              </>
-            )}
-          </AnimatePresence>
-        </ResizablePanelGroup>
-        {/* Dialogs */}
-        <DocumentDialog />
-        <CreateProjectDialog />
-      </Authenticated>
+      <motion.div
+        className="flex-1 min-h-0"
+        layout
+        transition={layoutTransition}
+      >
+        <ChatMessages />
+      </motion.div>
+      <motion.div className="flex-none" layout transition={layoutTransition}>
+        <ChatInput />
+      </motion.div>
     </>
   );
 }
