+++ src/store/chatStore.ts
@@ -31,12 +31,12 @@ export const sidebarOpenAtom = atomWithStorage("sidebarOpen", false);
 export const resizePanelOpenAtom = atomWithStorage("resizePanelOpen", false);
 export const selectedPanelTabAtom = atomWithStorage(
   "selectedPanelTab",
-  "projects",
+  "projects"
 );
 export const resizePanelWidthAtom = atomWithStorage("resizePanelWidth", 40);
 
 export const documentDialogOpenAtom = atom<Id<"documents"> | undefined>(
-  undefined,
+  undefined
 );
 export const createProjectDialogOpenAtom = atom(false);
 export const createMCPServerDialogOpenAtom = atom(false);
@@ -44,7 +44,7 @@ export const createMCPServerDialogOpenAtom = atom(false);
 export const wrapLongLinesAtom = atomWithStorage("wrapLongLines", false);
 
 export const selectedProjectIdAtom = atom<Id<"projects"> | undefined>(
-  undefined,
+  undefined
 );
 export const selectedArtifactAtom = atom<Artifact | undefined>(undefined);
 
@@ -52,15 +52,15 @@ export const groupedMessagesAtom = atom<
   ReturnType<typeof groupMessages> | undefined
 >(undefined);
 export const lastChatMessageAtom = atom<Id<"chatMessages"> | undefined>(
-  undefined,
+  undefined
 );
 
 export const useStreamAtom = atom<ReturnType<typeof useStream> | undefined>(
-  undefined,
+  undefined
 );
 export const streamStatusAtom = selectAtom(
   useStreamAtom,
-  (stream) => stream?.status,
+  (stream) => stream?.status
 );
 
 // Create a more stable derived atom
@@ -86,7 +86,7 @@ export const allArtifactsAtom = atom((get) => {
             const messageArtifacts = parts
               .filter(
                 (part): part is Extract<ContentPart, { type: "artifact" }> =>
-                  part.type === "artifact",
+                  part.type === "artifact"
               )
               .map((part) => part.artifact);
             artifacts.push(...messageArtifacts);
@@ -105,7 +105,7 @@ export const allArtifactsAtom = atom((get) => {
     const streamArtifacts = streamParts
       .filter(
         (part): part is Extract<ContentPart, { type: "artifact" }> =>
-          part.type === "artifact",
+          part.type === "artifact"
       )
       .map((part) => part.artifact);
